requirements.txt
asgiref==3.5.0
black==22.3.0
click==8.1.2
dj-database-url==0.5.0
Django==4.0.4
django-environ==0.8.1
django-heroku==0.3.1
django-nested-inline==0.4.5
djangorestframework==3.13.1
drf-writable-nested==0.6.3
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
whitenoise==6.0.0
